Revert "Add some tests to check `Display` impls"
This reverts commit 4a139744a54689a02d35f616620b29db121c254a.
